LayoutTests/canvas/philip/tests/2d.path.arc.negative-expected.txt

< [index] >
2d.path.arc.negative
arc() with negative radius throws INDEX_SIZE_ERR
References: 2d.path.arc.negative
Actual output:
Passed
